Efectos de fondo agregados, tabla score bug

+++ Assets/_scripts/HighScore/HighScoreManager.cs
@@ -117,10 +117,10 @@ public class HighScoreManager : MonoBehaviour
             _listScore.Add(new Score
             {
                 UserName = entryInfo[0],
-                Value = float.Parse(entryInfo[1])
+                Value = Int32.Parse(entryInfo[1])
             });
         }
-        
+
         //Ordenar la lista por puntaje
         _listScore.OrderBy(x => x.Value);
     }
